Move update code out of work_items folder and create cronjobs to call it
