+++ src/selectedtests/task_mappings/task_mappings_cli.py
@@ -8,9 +8,11 @@ from datetime import datetime
 import click
 import structlog
 
+from selectedtests.datasource.mongo_wrapper import MongoWrapper
 from selectedtests.helpers import get_evg_api
 from selectedtests.task_mappings.create_task_mappings import generate_task_mappings
 from selectedtests.task_mappings.version_limit import VersionLimit
+from selectedtests.task_mappings.update_task_mappings import update_task_mappings_since_last_commit
 
 LOGGER = structlog.get_logger(__name__)
 
@@ -28,8 +30,8 @@ def _setup_logging(verbose: bool):
 
 
 @click.group()
-@click.pass_context
 @click.option("--verbose", is_flag=True, default=False, help="Enable verbose logging.")
+@click.pass_context
 def cli(ctx, verbose: bool):
     """Entry point for the cli interface. It sets up the evg api instance and logging."""
     ctx.ensure_object(dict)
@@ -122,6 +124,19 @@ def create(
         print(json_dump)
 
     LOGGER.info("Finished processing task mappings")
+
+
+@cli.command()
+@click.option(
+    "--mongo-uri",
+    type=str,
+    default=lambda: os.environ.get("SELECTED_TESTS_MONGO_URI"),
+    help="Mongo URI to connect to.",
+)
+@click.pass_context
+def update(ctx, mongo_uri):
+    """Process task mappings since they were last processed."""
+    update_task_mappings_since_last_commit(ctx.obj["evg_api"], MongoWrapper.connect(mongo_uri))
 
 
 def main():
