+++ tests/selectedtests/task_mappings/test_task_mappings_cli.py
@@ -15,7 +15,7 @@ def ns(relative_name):
 class TestCli:
     @patch(ns("get_evg_api"))
     @patch(ns("generate_task_mappings"))
-    def test_arguments_passed_in(self, generate_task_mappings_mock, get_evg_api_mock):
+    def test_create_arguments_passed_in(self, generate_task_mappings_mock, get_evg_api_mock):
         mock_get_evg_api_mock = MagicMock()
         get_evg_api_mock.return_value = mock_get_evg_api_mock
         expected_mappings = ["mock-response"]
@@ -48,7 +48,7 @@ class TestCli:
 
     @patch(ns("get_evg_api"))
     @patch(ns("generate_task_mappings"))
-    def test_invalid_dates(self, generate_task_mappings_mock, get_evg_api_mock):
+    def test_create_with_invalid_dates(self, generate_task_mappings_mock, get_evg_api_mock):
         mock_get_evg_api_mock = MagicMock()
         get_evg_api_mock.return_value = mock_get_evg_api_mock
         generate_task_mappings_mock.return_value = ([], "most-recent-sha-analyzed")
@@ -80,7 +80,9 @@ class TestCli:
 
     @patch(ns("get_evg_api"))
     @patch(ns("generate_task_mappings"))
-    def test_module_regexes_not_passed_in(self, generate_task_mappings_mock, get_evg_api_mock):
+    def test_create_with_module_regexes_not_passed_in(
+        self, generate_task_mappings_mock, get_evg_api_mock
+    ):
         mock_get_evg_api_mock = MagicMock()
         get_evg_api_mock.return_value = mock_get_evg_api_mock
         generate_task_mappings_mock.return_value = ([], "most-recent-sha-analyzed")
@@ -108,3 +110,17 @@ class TestCli:
                 "A module source file regex is required when a module is being analyzed"
                 in result.stdout
             )
+
+    @patch(ns("get_evg_api"))
+    @patch(ns("MongoWrapper.connect"))
+    @patch(ns("update_task_mappings_since_last_commit"))
+    def test_update(
+        self, update_task_mappings_since_last_commit_mock, mongo_wrapper_mock, evg_api_mock
+    ):
+        evg_api_mock.get_api.return_value = MagicMock()
+        mongo_wrapper_mock.get_api.return_value = MagicMock()
+
+        runner = CliRunner()
+        with runner.isolated_filesystem():
+            result = runner.invoke(cli, ["update", "--mongo-uri=localhost"])
+            assert result.exit_code == 0
